+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 > The version number of fuji follows `semver` now: https://semver.org/
 
-- (command_bundle module) fix: the `/command-bundle un-register` command should not trigger NPE for some bundle-command.
+- (command_toolbox.tpa) feature: add `/tpaaccept all`, `/tpadeny all` and `/tpacancel all` commands for convenience. (Thanks to @NeverEndingCycle)